D合并后第五次提交

+++ kaohe/B/B.cpp
@@ -1,6 +1,6 @@
 #include "B.h"
 
-DLL_IMPORT_EXPORT	int sub_B(int a, int b)
+DLL_IMPORT_EXPORT	double sub_B(double a, double b)
 {
 	return a - b;
 }
